+++ src/main/java/de/htwsaar/prog3/carrental/controller/CustomerTableViewController.java
@@ -81,7 +81,7 @@ public class CustomerTableViewController extends GenericTableViewController<Cust
         Alert confirmationDialog = DialogUtil
                 .createConfirmationDialog(I18nComponentsUtil.getDialogDeleteConfirmationText());
         Optional<ButtonType> result = confirmationDialog.showAndWait();
-        if (result.get() == ButtonType.OK) {
+        if (result.isPresent() && result.get() == ButtonType.OK) {
             service.delete(toDelete);
             entities.remove(toDelete);
         }
